+++ acornui-core/src/jvmMain/kotlin/com/acornui/core/time/DateImpl.kt
@@ -1,32 +1,13 @@
-/*
- * Copyright 2015 Nicholas Bilyk
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+package com.acornui.core.time
 
-package com.acornui.jvm.time
-
-import com.acornui.core.time.Date
-import com.acornui.core.time.DateRo
 import com.acornui.core.zeroPadding
 import java.time.ZoneOffset
 import java.util.*
 
-
 /**
  * @author nbilyk
  */
-class DateImpl : Date {
+actual class DateImpl : Date {
 
 	private var localDateIsValid = true
 	private var utcDateIsValid = false
