+++ acornui-core/src/jvmTest/kotlin/com/acornui/core/serialization/BinarySerializerTest.kt
@@ -17,8 +17,10 @@
 package com.acornui.core.serialization
 
 import org.junit.Assert.*
+import org.junit.Test
 
 class BinarySerializerTest {
+
 
 
 }
