+++ acornui-utils/src/jvmMain/kotlin/com/acornui/core/debug.kt
@@ -1,0 +1,6 @@
+package com.acornui.core
+
+/**
+ * A flag for enabling various debugging features like debug logging.
+ */
+actual val debug: Boolean by lazy { System.getProperty("debug")?.toLowerCase() == "true" }
