+++ acornui-utils/src/jvmMain/kotlin/com/acornui/core/lineSeparator.kt
@@ -1,0 +1,7 @@
+package com.acornui.core
+
+/**
+ * May be set by the back-ends.
+ * JVM sets to System.lineSeparator()
+ */
+actual val lineSeparator: String = System.lineSeparator()
