+++ backends/acornui-lwjgl-backend/src/jvmTest/kotlin/com/acornui/jvm/time/DateParserTest.kt
@@ -4,12 +4,8 @@ import com.acornui.core.UserInfo
 import com.acornui.core.i18n.Locale
 import com.acornui.core.text.DateTimeFormatType
 import com.acornui.core.text.DateTimeParser
-import com.acornui.core.text.dateTimeFormatterProvider
-import com.acornui.core.text.numberFormatterProvider
 import com.acornui.core.time.time
 import com.acornui.core.userInfo
-import com.acornui.jvm.text.DateTimeFormatterImpl
-import com.acornui.jvm.text.NumberFormatterImpl
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -26,9 +22,7 @@ class DateParserTest {
 				systemLocale = listOf(Locale("en-US"))
 		)
 		userInfo = u
-		time = TimeProviderImpl()
-		numberFormatterProvider = { NumberFormatterImpl() }
-		dateTimeFormatterProvider = { DateTimeFormatterImpl() }
+
 	}
 
 	@Test
