+++ backends/acornui-lwjgl-backend/src/jvmTest/kotlin/com/acornui/jvm/time/DateTimeFormatterTest.kt
@@ -18,13 +18,8 @@ package com.acornui.jvm.time
 
 import com.acornui.core.UserInfo
 import com.acornui.core.i18n.Locale
-import com.acornui.core.text.dateTimeFormatterProvider
 import com.acornui.core.text.getMonths
-import com.acornui.core.text.numberFormatterProvider
-import com.acornui.core.time.time
 import com.acornui.core.userInfo
-import com.acornui.jvm.text.DateTimeFormatterImpl
-import com.acornui.jvm.text.NumberFormatterImpl
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -41,9 +36,6 @@ class DateTimeFormatterTest {
 				systemLocale = listOf(Locale("en-US"))
 		)
 		userInfo = u
-		time = TimeProviderImpl()
-		numberFormatterProvider = { NumberFormatterImpl() }
-		dateTimeFormatterProvider = { DateTimeFormatterImpl() }
 	}
 
 	@Test
