+++ backends/acornui-webgl-backend/src/jsMain/kotlin/com/acornui/js/text/NumberFormatterImpl.kt
@@ -25,52 +25,3 @@ import com.acornui.core.userInfo
 import com.acornui.reflect.observable
 import kotlin.properties.ReadWriteProperty
 
-class NumberFormatterImpl() : NumberFormatter {
-
-	override var type by watched(NumberFormatType.NUMBER)
-	override var locales: List<Locale>? by watched(null)
-	override var minIntegerDigits: Int by watched(1)
-	override var maxIntegerDigits: Int by watched(40)
-	override var minFractionDigits: Int by watched(0)
-	override var maxFractionDigits: Int by watched(3)
-	override var useGrouping: Boolean by watched(true)
-
-	override var currencyCode: String by watched("USD")
-
-	private var lastLocales: List<Locale> = listOf()
-	private var formatter: dynamic = null
-
-	override fun format(value: Number?): String {
-		if (value == null) return ""
-		if (locales == null && lastLocales != userInfo.currentLocale.value) {
-			formatter = null
-			lastLocales = userInfo.currentLocale.value.copy()
-		}
-		if (formatter == null) {
-			val locales = (locales ?: lastLocales).map { it.value }
-			val JsNumberFormat = js("Intl.NumberFormat")
-			val options = js("({})")
-
-			options.minimumIntegerDigits = minIntegerDigits
-			options.maximumIntegerDigits = maxIntegerDigits
-			options.minimumFractionDigits = minFractionDigits
-			options.maximumFractionDigits = maxFractionDigits
-			options.useGrouping = useGrouping
-			if (type == NumberFormatType.CURRENCY) {
-				options.style = "currency"
-				options.currency = currencyCode
-			} else if (type == NumberFormatType.PERCENT) {
-				options.style = "percent"
-			}
-
-			formatter = JsNumberFormat(locales.toTypedArray(), options)
-		}
-		return formatter!!.format(value)
-	}
-
-	private fun <T> watched(initial: T): ReadWriteProperty<Any?, T> {
-		return observable(initial) {
-			formatter = null
-		}
-	}
-}
